\name{Initialization}
\alias{Initialization}
\title{
Create all formatting and variable initialization for the test and report creation process
}
\description{
This is a pseudo-function with no input that declares all variables and pastes all formatting options for the test/report creation process to the main document.
}
\usage{
Initialization(Initialize = TRUE, Version)
}
\arguments{
  \item{Initialize}{
    A pseudo-variable. It is not meant to be changed or declared.
}
  \item{Version}{ The run version. }
}
\value{
Returns all necessary variables and formatting options for the test and report creation process
}
\author{
Adam Rahman
}
