\name{Initialization}
\alias{Initialization}
\title{
Create all formatting and variable initialization for the test and report creation process
}
\description{
This function declares all variables and pastes all formatting options for the test/report creation process to the main document.
}
\usage{
Initialization(name, version = Version())
}
\value{
Returns all necessary variables and formatting options for the test and report creation process
}
\author{
Adam Rahman
}
